Format the code.

+++ src/Network/Socket.cpp
@@ -503,12 +503,12 @@ uint64_t Socket::elapsedTimeAfterFlushed() {
     return _send_flush_ticker.elapsedTime();
 }
 
-int Socket::getRecvSpeed() {
+size_t Socket::getRecvSpeed() {
     _enable_speed = true;
     return _recv_speed.getSpeed();
 }
 
-int Socket::getSendSpeed() {
+size_t Socket::getSendSpeed() {
     _enable_speed = true;
     return _send_speed.getSpeed();
 }
